+++ src/components/CheckinDetail.component.tsx
@@ -3,14 +3,14 @@ import { View, Text, StyleSheet, FlatList } from 'react-native'
 import { Avatar, Image, Icon } from 'react-native-elements'
 import { NavigationProp } from '@react-navigation/native'
 import { ImageCarousel } from '@/components/carousel/ImageCarousel.component'
-import window from '../constants/Layout'
-import { useDate } from '../hooks/useDate'
-import { useUtils } from '../hooks/useUtils'
-import colors from '../constants/Colors'
-import { useFoursquare } from '../hooks/useFoursquare'
-import { Checkin, User } from '../types/Foursquare'
-import { commonStyles } from '../styles/styles'
-import CoinIcon from './CoinIcon.component'
+import { useDate } from '@/hooks/useDate'
+import { useUtils } from '@/hooks/useUtils'
+import { useFoursquare } from '@/hooks/useFoursquare'
+import colors from '@/constants/Colors'
+import window from '@/constants/Layout'
+import { Checkin, User } from '@/types/Foursquare'
+import { commonStyles } from '@/styles/styles'
+import CoinIcon from '@/components/CoinIcon.component'
 
 export const CheckinDetail = ({ route, navigation }) => {
   const { item }: { item: Checkin } = route.params
