+++ src/components/CoinIcon.component.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { View } from 'react-native'
-import colors from '../constants/Colors'
+import colors from '@/constants/Colors'
 const CoinIcon = () => {
   return (
     <View
