+++ src/components/EditScreenInfo.tsx
@@ -2,9 +2,9 @@ import * as WebBrowser from 'expo-web-browser'
 import React from 'react'
 import { StyleSheet, TouchableOpacity } from 'react-native'
 
-import Colors from '../constants/Colors'
-import { MonoText } from './StyledText'
-import { Text, View } from './Themed'
+import Colors from '@/constants/Colors'
+import { Text, View } from '@/components/Themed'
+import { MonoText } from '@/components/StyledText'
 
 export default function EditScreenInfo({ path }: { path: string }) {
   return (
