+++ src/components/StyledText.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import { Text, TextProps } from './Themed'
+import { Text, TextProps } from '@/components/Themed'
 
 export function MonoText(props: TextProps) {
   return <Text {...props} style={[props.style, { fontFamily: 'space-mono' }]} />
