+++ src/components/Themed.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import { Text as DefaultText, View as DefaultView } from 'react-native'
 
-import Colors from '../constants/Colors'
-import useColorScheme from '../hooks/useColorScheme'
+import useColorScheme from '@/hooks/useColorScheme'
+import Colors from '@/constants/Colors'
 
 export function useThemeColor(
   props: { light?: string; dark?: string },
