+++ src/components/__tests__/StyledText-test.js
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import renderer from 'react-test-renderer'
 
-import { MonoText } from '../StyledText'
+import { MonoText } from '@/components/StyledText'
 
 it(`renders correctly`, () => {
   const tree = renderer.create(<MonoText>Snapshot test!</MonoText>).toJSON()
