+++ src/components/card/checkin.component.tsx
@@ -4,13 +4,13 @@ import { Avatar, Image, Icon } from 'react-native-elements'
 import { ScrollView } from 'react-native-gesture-handler'
 import ImageViewer from 'react-native-image-zoom-viewer'
 import { useNavigation } from '@react-navigation/core'
-import type { Checkin } from '../../types/Foursquare'
-import window from '../../constants/Layout'
-import { useDate } from '../../hooks/useDate'
-import { useUtils } from '../../hooks/useUtils'
-import colors from '../../constants/Colors'
-import { useRecoil } from '../../hooks/useRecoil'
-import { commonStyles } from '../../styles/styles'
+import type { Checkin } from '@/types/Foursquare'
+import window from '@/constants/Layout'
+import { useDate } from '@/hooks/useDate'
+import { useUtils } from '@/hooks/useUtils'
+import colors from '@/constants/Colors'
+import { useRecoil } from '@/hooks/useRecoil'
+import { commonStyles } from '@/styles/styles'
 
 export const CheckinCard = ({ item }: { item: Checkin }) => {
   const navigation = useNavigation()
