+++ src/navigation/index.tsx
@@ -3,9 +3,9 @@ import { createStackNavigator } from '@react-navigation/stack'
 import * as React from 'react'
 import { ColorSchemeName } from 'react-native'
 
-import { RootStackParamList } from '../types'
-import BottomTabNavigator from './BottomTabNavigator'
-import LinkingConfiguration from './LinkingConfiguration'
+import { RootStackParamList } from '@/types'
+import BottomTabNavigator from '@/navigation/BottomTabNavigator'
+import LinkingConfiguration from '@/navigation/LinkingConfiguration'
 
 // If you are not familiar with React Navigation, we recommend going through the
 // "Fundamentals" guide: https://reactnavigation.org/docs/getting-started
