+++ src/styles/styles.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native'
-import colors from '../constants/Colors'
+import colors from '@/constants/Colors'
 
 export const commonStyles = StyleSheet.create({
   bk_white: { backgroundColor: colors.light.background, height: '100%' },
